Moved State from Line Graph component to App
